requirements.txt
colorama==0.4.6
grpcio==1.58.0
grpcio-tools==1.58.0
protobuf==4.24.3
